Split up ptr/mod.rs in libcore, one with implementation detail for const ptr and the other with mut ptr
